requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
executing==2.2.1
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
limits==5.8.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
ordered-set==4.1.0
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
six==1.17.0
SQLAlchemy==2.0.48
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.2
WTForms==3.2.1
